Make it clear that `bind_definition` does not need to return KeyAnnotation
Summary:
I'm working on implementing basic support for function decorators,
and one of the things I ran into is that I need access to the bound
key coming from the function definition.

This isn't very convenient right now, because `bind_definition` already returns
an optional `KeyAnnotation`, which means the return type would become kind of
complicated.

But I realized skimming the code that this is actually never needed:
`bind_defintion` is *exclusively* used for side effects right now. This diff
tweaks the type signature to make that obvious.

(Aside: why was it returning Annotation at all? It used to be the only bind
operation, and we need it for some use cases. But all those use cases - for
example validating assignments in `bind_target` - are all now using the lower
level `bind_key` which means we no longer need it.)

Reviewed By: ndmitchell

Differential Revision: D68208403

fbshipit-source-id: 6a3e9413d9542265d9d55a86248e36e30302ab77