Construct Key::ExpectedReturnType and Binding::GeneratorReturnType for use in return statement validation and inference
Summary:
Following this doc https://docs.google.com/document/d/1CZYhG-zDJ5hkbxk08xOetsDFAlkG0umShNVmiIDpZfM/edit?tab=t.0

We are creating a new key and a new binding to be used in return statement validation and inference.

- `ExpectedReturnType` key, which will be used to represent the return type after considering generators (as opposed to simply the type of the return if the function).  For this, I created a similar data structure to that as `ReturnType`

- `GeneratorReturnType` binding which we can use to add the existing annotation and decompose later so we can do compatability checks against the return type as opposed to the whole generator type. For this, I made the structure similar to that of the AnnotatedType binding but made the key an optional annotation `Option<Idx<KeyAnnotation>>`

Looking for comments on the structures or any suggestions for improvements. I will use these structures in answers. Currently this is marked as a todo.

Reviewed By: stroxler

Differential Revision: D68240365

fbshipit-source-id: 6179548874f74c2bf5e0e0ec11862de10efdb286