Add a (failing) test of decorator ordering
Summary:
We had tests of simple decorators, signatue-preserving generic decorators,
and signature-modifying decorators.

But we also should verify that we correctly evaluate multiple decorators,
and we do so in the correct order. In this case `decorator0` and `decorator1`
wrap the output in `set` and `list` respectively, so we can verify that
they both are evaluated and in what order.

The test fails right now because we're processing decorators in top-down
order whereas they should be bottom-up; I'll fix this in the next diff.

Reviewed By: ndmitchell, grievejia

Differential Revision: D68247507

fbshipit-source-id: 8457a42b028695312d2d5e3a91a73decbe489c80