Fix initialize-to-`...`-in-stub behavior
Summary:
This commit separates Pyre2's decision about whether to type check the
value in an `AnnAssign` versus whether to treat it as initialized.

We need this so that we can correctly skip typing `...` assignment in stubs,
while also correctly treating it as indicating initialization.

This fixes a bug in class-and-instance attribute handling from stubs.

Reviewed By: rchen152

Differential Revision: D68279121

fbshipit-source-id: f0a896e86b1bd84c392566bd7e81ca255aa71c87