Restructure AnnAssign a bit
Summary:
I need to make a behavioral change to this logic to fix class
visibility rules for attributes defined in stubs with ellipses.

But I realized that with the current code structure, my change is
pretty hard to read because I have to do a lot of refactoring
and modify behavior at the same time.

The underlying problem is partially that the existing code structure
isn't very DRY: we're doing the entire bind operation in a branch,
which is share-nothing, but it's really only the binding and initialization
tracking that differ.

This commit makes that clearer by defining the binding and initialized
flag in a branch but only calling `bind_definition` once at the end.

Reviewed By: rchen152

Differential Revision: D68279118

fbshipit-source-id: 6f41ee56b83cb28c835c94daf43b4fe888f5e786