Delete confusing FIXME
Summary: Not entirely sure what this means. I think it's hinting that would could produce the module directly with a Type::Module. And I think we could. But it's not clear that's a better thing to do. We may want to simplify this and do a bit more in bindings, but no need for now.

Reviewed By: yangdanny97

Differential Revision: D68440438

fbshipit-source-id: 2cf68eebafb403aca8d7a8b8d35d6f7ca11c5bbf