Apply and ignore class decorators
Summary:
Plumbs class decorators through bindings.rs so we can start supporting `dataclass`. Ignores them for now in expr.rs so we don't start generating tons of errors.

I also had to make a slightly weird change to check for `builtins.type` by name in a bit of code that now runs when the stdlib is being bootstrapped.

Reviewed By: stroxler

Differential Revision: D68465433

fbshipit-source-id: 7815e06f816c0b802a6a5c065abc87250c04bb43