enum functional syntax
Summary:
This diff adds support for functional syntax for declaring enums:

https://typing.readthedocs.io/en/latest/spec/enums.html#enum-definition

The other forms will be added in later diffs, and the overall setup should be useful reference for other functional decls (NamedTuple, TypedDict, etc).

I also add a util function that checks whether a string is a valid identifier. note that this doesn't check for keywords yet.

Reviewed By: stroxler, ndmitchell

Differential Revision: D67879720

fbshipit-source-id: 72c25dec1b4494fb0ece2debf48e109f3de1c521