start implementing protocol typechecking
Summary:
This diff implements the following for protocols:
- If the base class list contains `Protocol` then the class is a protocol, and all other base classes must be protocols
- Protocols are not assignable to concrete classes
- Basic structural checking, as outlined below.

Structural checking currently just checks that an attribute with that name is present, without checking the type, similar to how `runtime_checkable` functions.

My naive attempt to check the types results in a stack overflow, since classes can be recursive. This makes me think that we should keep a stack of types being checked in the subset function, in addition to gas. If we're checking X1 <: X2 and a few layers deeper we encounter another X1 <: X2, I think we should assume the latter is true, based on the typing spec:

> for recursive protocols, a class is considered assignable to the protocol in situations where the decision depends on itself

Reviewed By: ndmitchell

Differential Revision: D68605863

fbshipit-source-id: 176a90307432c27e00774b5be4bcb7046b15f25a