Allow `from PACKAGE import MODULE`
Summary:
moarpackages

Previously, when seeing the following statement:
```
from a.b import c
```
Pyre2 would only look up the content of module `a.b`, and see if there's a name `c` defined there.

But in Python, the name `a.b` may also point to a package -- Either there's an explicit `a/b/__init__.py` (normal package), or there's no `__init__.py` but some other python files just exist under `a/b/` (namespace package). In those cases, the from statement would bind the name `c` to the module `a/b/c.py`, even if the name `c` is not defined anywhere in `a/b/__init__.py` (if there exists one)!

This diff adds support for this kind of package-module import, by checking for module existence of `a.b.c` when `c` cannot be found defined in `a.b`. I think this would *mostly* cover the case -- there might be some prioritization issues that I might dig into later (e.g. when `c` is defined both in __init__ and there's a module with that name). But this diff at least could server as a reasonable starting point, and help us get rid of the `from importlib import _bootstrap_external` issue in typeshed.

Reviewed By: ndmitchell

Differential Revision: D68599507

fbshipit-source-id: 875beac1ba2b1750daf2694ad97ee3205a8fb0cf