Eliminate direct use of metaclass iterate (regression)
Summary:
This diff may be a bit controversial, it's a temporary step backward in end-user
behavior but I'm confident it is a step in the right direction.

In order to rework attribute access to involve "get" and "set" actions without
exacerbating existing tech debt, I need our code to respect layering better, and
the `self.call_metaclass_method` function in `expr.rs` is reaching too far
into the guts of `classes.rs` (and inappropriately hardcoding unusual
ignore-anything-from-`type` logic that is actually wrong for every case except
constructors).

The current handling of `iterate` is relying on this logic that needs to go.
For the most part, this isn't necessary - in every case where Pyre2 is capable
of correctly analyzing actual calls, we can use normal attribute lookups, which
is what this diff does (in the process eliminating significant complexity
and solving the modularity issues that are in the way of my attribute action changes
that block `property` and `classmethod`).

But handling `EnumMeta`'s `__iter__` method is beyond Pyre2's current capabilities:
we incorrectly turn methods into plain callable types before calling them most
of the time. This will not work in the long run, because annotations on `self`
can be used to define overloaded behavior, or to rule out calling methods on specific
types.

For example, it is entirely possible for a user to define a class like this
```
from typing import overload

class A[T]:
     overload
     def foo(self: A[int]) -> str: ...

     overload
     def foo(self: A[str]) -> int: ...

```
in which case `a.foo` takes `A[int]` to `str`, `A[str]` to `int`, and is a type
error for any incompatible generic instance of `A`. This may seem like an edge
case but it does happen - for example essentially *all* of `PyreReadOnly` relies
on exactly this sort of thing.

The issue with `EnumMeta` has the same root cause: it uses an annotation on `self`
to let the constraint solver to tie the input type to the result, which requires
modeling the call of the bound method as a full entity rather than trying to
eagerly extract a callable type with the `self` type stripped.

We were emulating this behavior, but *only* for metaclass calls, and using that
to make this test case pass. But we really need to achieve the same end by making
all of Pyre2 handle calls correctly rather than hacking a metaclass-only fix. As
a result, I'd like to regress this test case with a clear TODO, so that I can
move forward with the changes I need for descriptors like `protocol` and
`classmethod`

Reviewed By: rchen152

Differential Revision: D68639580

fbshipit-source-id: a3c512a283fe6dcee0e465806c82107998f9b5cb