Infer YieldFrom expression type
Summary:
In this diff, we propagate the return type of a generator to expr_infer.

When creating bindings, we not only add hooks to the send type but to the return type.

When we reach expr infer and encounter a YieldFrom, we extract the return type.

RFC:

in the unittest:

def my_generator_nested() -> Generator[Yield, Send, Return]:
    yield Yield()
    return Return()

def my_generator() -> Generator[Yield, Send, Return]:
    s = yield Yield()
    y = yield from  my_generator_nested()

    reveal_type(s) # E: revealed type: Send
    reveal_type(y) # E: revealed type: Return

    return Return()

we have two occurrences of the same annotation `Generator[Yield, Send, Return]`.  In the end, they do need to be compatible, so the behavior right now is correct.

Right now, I am adding a hook to the annotation associated with `my_generator`.

Do we also need to add a hook to `my_generator_nested`?

I will look into that next.

Reviewed By: stroxler

Differential Revision: D68643899

fbshipit-source-id: 2444e53645328317abcb59931966a5d51a0674ad