Create an `Attribute` type and re-wire access logic
Summary:
This commit begins cutting `attr.rs` over to an attribute-action oriented
interface, by eliminating `AccessNotAllowed` as a top-level variant of
`LookupResult` and instead modeling the lookup result as a `Attribute`,
which can in turn be queried for an `AttributeAccess` that knows whether
the access is permitted (and if so the type, if not the reason why not).

We don't do anything new with this capabiliity yet (for now we're just
pushing around where the no-access-allowed-on-class-object information
gets stored), but with this change we are much closer to being able to
model descritptors and other access-rule-based features like `Final`.

Reviewed By: ndmitchell

Differential Revision: D68687293

fbshipit-source-id: 75ba457a5a3ad00191b8ca1e9699bd5efe9fd428