Make a new `AttributeAccess` data structure
Summary:
Eventually this data structure will define get and set actions, which either
are permitted with some `Type`, or are not permitted and have a reason.

For now though, we just use this to greatly simplify the bridge between
`get_class_attribute` and `attr.rs` (this bridge has been a pain point
for a long term, because we were duplicating error reasons and having
to cast between them).

The new interface returns an optional `AttributeAccess`: if the result is
None then the attribute was not found (which in many cases causes us to
then attempt an instance lookup on the metaclass), whereas if the result
is `Some` then we either have a type or the access is not allowed.

This representation is much more aligned with what we actually need than
the previous representation where "not found" was still modeled as an
error branch.

Reviewed By: ndmitchell

Differential Revision: D68687294

fbshipit-source-id: 30272e4168d2eddf9a7fccf59faa7108d4a465ff