Add a test of a type alias parametrized by a generic class
Summary:
The question of whether we allow type aliases inside classes has come up
a few times, and for "flat" aliases that don't reference a class's generic
type parameters the answer is an unqualified (and I think uncontroversial)
yes - you can namespace a plain alias inside a class, no problem.

More controversial is the question of what happens if the alias tries to
reference class type parameters.

Regardless of what we think *should* happen here, it's good to have a test
demonstrating what actually does happen:
- We do allow a type alias that uses a class type parameter to be nested
  inside a class. It can be used freely directly in the class body, or on
  instances of the class.
- We do *not* allow it to be accessed from the class directly, even for
  an instantiated version of the class (e.g. `C[int].X` in the new test
  case). The reason is that statically typing a *value* of a class object whose
  type is generic doesn't really make sense. This is arguably the wrong
  behavior for a type alias; we could add a special case to allow it.

Opinions are welcome, I can add a TODO for this if we think `C[int].X`
should be legal.

Reviewed By: ndmitchell

Differential Revision: D68662736

fbshipit-source-id: fd48e44f6c8d87c03513c20f1a777661468c4461