Move decorator application inside function/class binding
Summary:
Currently, decorators wrap function/class keys and also overwrite the relevant entry in the environment.

For overloads, we want to augment the environment a bit, because a given function name can point to more than one definition.

Because the env is a map with a string key, I am going to add a new bind_function_definition function, which _appends_ to the value for a given key instead of overwrites.

Moving decorator application in this way makes that change simpler. I probably wouldn't change where decorators are applied otherwise, but I also think that the proposed change is nice on its own. We avoid the additional key/binding pair, and need to do less juggling with Idx and env stuff.

Reviewed By: stroxler

Differential Revision: D68657220

fbshipit-source-id: 1250f51fc6a61c3992f55069826894de7a182ecd