Change Visitors::stmts callback to take a block instead of individual statements
Summary:
For overloads, I want to detect a sequence of function definitions statements with the same name. The simplest place to do this is while iterative over statements.

The visitor is used in Definitions::stmt, and I want to make sure the in Definition::stmts is used.

For now, the stmts function is just a simple iteration over the statements, but in later diffs I will add logic to combine a sequence of definitions into a single definition.

(I also think it's a bit nice to convert the iterator chaining into more direct-style/imperative code, but that's just a personal preference.)

Reviewed By: grievejia

Differential Revision: D68740981

fbshipit-source-id: 786bda7b821b4fab7fee6f889bc16a035e3522b6