Wire up CheckAssignTypeToAttribute for Assign
Summary:
This requires some careful plumbing inside of `bind_target`:
- We rely on `make_binding` to handle the inference case, which is used to
	provide a fallback binding we might use for inferring the attribute type if
	we never find an explicit annotation.
- We directly create the check binding itself, which relies on the computed
	type that might or might not come from ^^ depending on whether we find an
	annotation elsewhere, inside of `bind_target`.

Note that this simple logic does not attempt to use contextual typing; doing
that is more difficult because we have to decide when the RHS can be decomposed
at an expression level (in which case we can contextually type the sub-expressions
associated with individual assignment targets) and when it cannot (in which
case we can only use non-bidirectional type-level analysis). Since the type-level
analysis is the more general case, for now we do only that.

Note that the code for the attribute case of `make_target` is now considerably more
complicated because we want to contextually type when we can, so we have to switch
on whether a value expression is available or not. The same complexity eventually needs
to be handled for the other cases as well. The general solution will probably look
a little bit different: the attribute case is a little unusual in that we're
happy to call `make_binding` with no annotation to produce the binding for the
self-assignment since that binding only gets evaluated if there is no
annotation for the attribute anyway.

The broken test case where unpacked attribute assigns were never checked is now
fixed (the similar case assigning to annotated names remains broken).

Reviewed By: ndmitchell

Differential Revision: D68767618

fbshipit-source-id: e4ea77b0043ca1ebca8d06f0f6d95ae8d7814ce6