Extract function_def
Summary:
Context:
For overloads, we will collect a sequence of function definitions with the same name into a single binding. In the answers phase, we can determine if that sequence defines an overloaded function and create the appropriate type.

This diff:
To prepare for a single Binding::Function with multiple FunctionBinding values, extract a function that we can call in a loop.

Reviewed By: ndmitchell

Differential Revision: D68800873

fbshipit-source-id: 1ff54071146eb4001946b6355526493a1ddc761c