Add test showing current definition count for overloaded functions
Summary:
Context:
For overloads, we will collect a sequence of function definitions with the same name into a single binding. In the answers phase, we can determine if that sequence defines an overloaded function and create the appropriate type.

This diff:
The definitions phase determines the number of bindings for a given value into the `count` field. This value should match the number of times we call `bind_key`, otherwise we will get weird behavior.

This diff just adds a test to show the current counts for sequences of function definitions. The next diff will change it.

Reviewed By: ndmitchell, rchen152

Differential Revision: D68800876

fbshipit-source-id: 408e854a1e570cfd3d83d7ea81caa0590c38d161