Buffer function definitions with the same name in bindings
Summary:
Context:
For overloads, we will collect a sequence of function definitions with the same name into a single binding. In the answers phase, we can determine if that sequence defines an overloaded function and create the appropriate type.

This diff:
In order to create a single binding for the sequence, we need to buffer the together. This diff is just the buffering code. We will still create a separate binding for each definition.

Reviewed By: ndmitchell

Differential Revision: D68800874

fbshipit-source-id: 0337f55b4fb4914f84552e02d2420453d5807466