Lock in the enum `_value_` lookup
Summary:
The old `type_of_attr_get_if_gettable` violates normal
rules about how attribute lookups should work, and is impossible
to support easily in the v0 property implementation I want.

But it is only used for exactly one thing: looking up the
`_value_` attribute of an enum, which is supposed to always be
a regular class field annotation. This means that the access is
always `ReadWrite`, so we can support it just fine if we push down
the knowledge of the actual use case into `attr.rs`.

As an added benefit, `calculate_class_field` (which is complex enough
to be a little hard to skim) gets a bit simpler :)

Reviewed By: ndmitchell

Differential Revision: D68897127

fbshipit-source-id: d0504458422da036c0cb9a428277d7f89dea8079