Add a test that we can use class body scope names
Summary:
I'm not entirely sure why the use of property setters leads to name
errors in decorations; I wrote this test case to rule out one possible
explanation, that class body is preventing name lookups (which tells
me that something in `function_def` is probably to blame).

The test seems like a good one to have though, and as far as I can tell
no existing test verified that we could reuse names bound in the flow
environment, which we should be able to.

Reviewed By: rchen152

Differential Revision: D68898143

fbshipit-source-id: 72525e6565e622b0f972b70bcd14d85c6faf1c75