Use a dummy method to break def folding for preperty setter test
Summary:
We discovered today that some def-folding logic added to support `overload` has
a side effect of making a given name bound to a function def undefined when analyzing
the next function def if the second def has the same name.

This doesn't work in general: there are several legitimate ways to use the fact
that at runtime the first def is defined, such as using the first def in a default
parameter or referring to it in a decorator.

In particular, this directly impacts property setter decorators, which are invoking
a method defined on the (read-only) property returned by the original `property`
decorator.

In order to avoid property setter support getting in the way of in-progress overloads,
for now I'm using a dummy method to break this folding logic. There's a TODO to
remove this restriction as the overload implementation progresses.

Reviewed By: ndmitchell

Differential Revision: D68910940

fbshipit-source-id: b9e856729a6eff7f09b2f49f3561e1edfc1f59b5