Infer yield type for async generators
Summary:
Propagate the information we get from the AsyncGenerator type annotation to infer the yield type.

TODO:
I think we need to add an extra field to the binding where we decompose the generator type. The field would indicate whether the function is async so we can check against the annotation we extract.

In the same place we are decomposing those types, we can also check that field and raise errors accordingly.

I will do this in the next diff unless there are any other thoughts.

Reviewed By: stroxler

Differential Revision: D68917451

fbshipit-source-id: abb6f7251dd74dc16fc12d2569db8e5e2fde6c39