Add flag to indicate if a function is async and use to check against annotations
Summary:
As per our previous diff, we add a flag to generator type annotations where we extract yield exoression types so that on yield expressions, we can check the annotation against the flag and raise an error accordingly.

By looking at pyright, I see that the behavior is symetric for Generators/AsyncGenerators with respect to the async flag.

Reviewed By: stroxler

Differential Revision: D68926177

fbshipit-source-id: 896eeb1d2c598e5bdfef275fb09d54ef114f9fc0