support enum.member, enum.nonmember & more
Summary:
This diff adds support for the `member` decorator, and initial values wrapped with `nonmember()`, `staticmethod()`, and `classmethod()`.

I'm not sure if the opposite (`nonmember` and `member()`) are useful since methods are non-members by default, and unwrapped values are members by default.

This diff also adds a special case for `enum.property`, aka `_magic_enum_attr` in Python 3.11+. With this, there are no more typeshed/stdlib/enum.pyi errors when checking the empty file.

My understanding of decorators is limited, but it does seem odd that unless it's specially handled as a Decorator variant there's no way of differentiating between a non-callable and a decorated callable.

Reviewed By: ndmitchell

Differential Revision: D69096190

fbshipit-source-id: 7494fa1d21bc2203bcde374031df96cbf47e1919