Add failing test for multiple annotations
Summary:
My thought for how to resolve: If a binding has an annotation anywhere, we should choose the "first" one and require that any other annotations are equal to that.

For "first" I'm open to various definitions, but I think any definition is ok. The simplest definition to implement is by first occurrence in the file.

We can then move this check out of the flow merge code and into bind_key.

Reviewed By: ndmitchell

Differential Revision: D69147454

fbshipit-source-id: 6987b94d85c7bdcf07643b9cc72f9d3eef563f0b